+++ docs/changes.rst
@@ -4,7 +4,15 @@ Changelog
 in development
 --------------
 
-* new settings for customisations
+* DataFile size is now a BigInteger field
+* New settings for customisations, contextual view overrides (eg INDEX_VIEWS).
+* A new AbstractTardisAppConfig class that all new tardis apps should subclass
+* Third-party tardis app dependency checking
+* Removed database index from Parameter.string_value to allow longer strings in
+  Postgres. Migrations add a Postgres partial index for string_values shorter
+  than 256 characters.
+* Changed constraints on the instrument model; facility and instrument name are
+  now unique together
 
 
 3.6 - 16 March 2015